Cleanup and fixed tools being interupted when charging
