+++ enderio-armory/src/main/java/com/enderio/armory/api/capability/IDarkSteelCapability.java
@@ -4,19 +4,21 @@ import java.util.Collection;
 import java.util.Optional;
 
 public interface IDarkSteelCapability {
+
     void addUpgrade(IDarkSteelUpgrade upgrade);
 
     void removeUpgrade(String name);
 
+    boolean hasUpgrade(String upgrade);
+
     boolean canApplyUpgrade(IDarkSteelUpgrade upgrade);
-
-    <T extends IDarkSteelUpgrade> Optional<T> getUpgradeAs(String upgradeName, Class<T> as);
-
-    Optional<IDarkSteelUpgrade> getUpgrade(String upgrade);
 
     Collection<IDarkSteelUpgrade> getUpgrades();
 
-    boolean hasUpgrade(String upgrade);
+    Optional<IDarkSteelUpgrade> getUpgrade(String upgrade);
+
+    <T extends IDarkSteelUpgrade> Optional<T> getUpgradeAs(String upgradeName, Class<T> as);
 
     Collection<IDarkSteelUpgrade> getUpgradesApplicable();
+
 }
