+++ enderio-armory/src/main/java/com/enderio/armory/client/EnderIOArmoryClient.java
@@ -2,7 +2,9 @@ package com.enderio.armory.client;
 
 import com.enderio.armory.EnderIOArmory;
 import com.enderio.armory.client.renderer.ElytraUpgradeRenderLayer;
+import com.enderio.armory.client.renderer.MultiEnergyBarDecorator;
 import com.enderio.armory.client.renderer.SolarUpgradeRenderLayer;
+import com.enderio.armory.common.init.ArmoryItems;
 import net.minecraft.client.renderer.entity.player.PlayerRenderer;
 import net.neoforged.api.distmarker.Dist;
 import net.neoforged.bus.api.SubscribeEvent;
@@ -10,6 +12,7 @@ import net.neoforged.fml.ModContainer;
 import net.neoforged.fml.common.EventBusSubscriber;
 import net.neoforged.fml.common.Mod;
 import net.neoforged.neoforge.client.event.EntityRenderersEvent;
+import net.neoforged.neoforge.client.event.RegisterItemDecorationsEvent;
 import net.neoforged.neoforge.client.gui.ConfigurationScreen;
 import net.neoforged.neoforge.client.gui.IConfigScreenFactory;
 
@@ -30,6 +33,17 @@ public class EnderIOArmoryClient {
                 playerRenderer.addLayer(new SolarUpgradeRenderLayer(playerRenderer));
             }
         }
+    }
+
+    @SubscribeEvent
+    public static void registerItemDecorations(RegisterItemDecorationsEvent event) {
+        event.register(ArmoryItems.DARK_STEEL_SWORD, MultiEnergyBarDecorator.INSTANCE);
+        event.register(ArmoryItems.DARK_STEEL_AXE, MultiEnergyBarDecorator.INSTANCE);
+        event.register(ArmoryItems.DARK_STEEL_PICKAXE, MultiEnergyBarDecorator.INSTANCE);
+        event.register(ArmoryItems.DARK_STEEL_HELMET, MultiEnergyBarDecorator.INSTANCE);
+        event.register(ArmoryItems.DARK_STEEL_CHESTPLATE, MultiEnergyBarDecorator.INSTANCE);
+        event.register(ArmoryItems.DARK_STEEL_LEGGINGS, MultiEnergyBarDecorator.INSTANCE);
+        event.register(ArmoryItems.DARK_STEEL_BOOTS, MultiEnergyBarDecorator.INSTANCE);
     }
 
 }
