+++ enderio-armory/src/main/java/com/enderio/armory/client/renderer/GliderIntegrationClient.java
@@ -1,4 +1,4 @@
-package com.enderio.armory.client;
+package com.enderio.armory.client.renderer;
 
 import com.enderio.base.api.integration.ClientIntegration;
 import com.enderio.base.client.EnderIOBaseClient;
